Header + footer + cleaning up
Here is the header with text with that has a hover affect. Also added a footer that is green with nothing else in it. The header and footer is also combined with the other content. Now the website containts of a header a red slide witha picture than a anaother sldie with cards and a footer