Added an order confirmation page
The program now includes an order confirmation page. After the user clicks "Submit" on the BurgerOrderer.html page, they'll be directed to this new page. Here, a thank you message will be displayed, along with a box showing an image of the ordered burger and another box listing the ingredients and their quantities. Users can also return to the main page by clicking the "Our Burgers" button.